Integrate time

+++ api/isxImage.h
@@ -28,7 +28,7 @@ public:
         , m_numChannels(inNumChannels)
     {
         assert(inWidth > 0 && inHeight > 0 && inRowBytes > 0 && inNumChannels > 0);
-        assert(m_rowBytes >= m_width * getPixelSizeInBytes());
+        assert(size_t(m_rowBytes) >= m_width * getPixelSizeInBytes());
         m_pixels.resize(getImageSizeInBytes());
     }
 
@@ -43,7 +43,7 @@ public:
         , m_timeStamp(inTimeStamp)
     {
         assert(inWidth > 0 && inHeight > 0 && inRowBytes > 0 && inNumChannels > 0);
-        assert(m_rowBytes >= m_width * getPixelSizeInBytes());
+        assert(size_t(m_rowBytes) >= m_width * getPixelSizeInBytes());
         m_pixels.resize(getImageSizeInBytes());
     }
